requirements.txt
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
adal==1.2.4
applicationinsights==0.11.9
argon2-cffi==20.1.0
attrs==19.3.0
azure-common==1.1.25
azure-core==1.7.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.1.0
azureml-automl-core==1.11.0
azureml-automl-runtime==1.11.0
azureml-core==1.11.0
azureml-dataprep==2.0.2
azureml-dataprep-native==14.2.1
azureml-dataset-runtime==1.11.0.post1
azureml-defaults==1.11.0
azureml-explain-model==1.11.0
azureml-interpret==1.11.0
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.11.0
azureml-pipeline-core==1.11.0
azureml-pipeline-steps==1.11.0
azureml-sdk==1.11.0
azureml-telemetry==1.11.0
azureml-train==1.11.0
azureml-train-automl==1.11.0
azureml-train-automl-client==1.11.0
azureml-train-automl-runtime==1.11.0.post1
azureml-train-core==1.11.0
azureml-train-restclients-hyperdrive==1.11.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.1.5
boto==2.49.0
boto3==1.14.35
botocore==1.17.35
catboost==0.24
category-encoders==2.2.2
#certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
configparser==3.7.4
contextlib2==0.6.0.post1
cryptography==3.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distro==1.5.0
docker==4.2.2
docutils==0.15.2
dotnetcore2==2.1.14
entrypoints==0.3
fire==0.3.1
Flask==1.0.3
fusepy==3.0.1
gensim==3.8.3
graphviz==0.14.1
gunicorn==19.9.0
idna==2.10
importlib-metadata==1.7.0
interpret-community==0.14.1
interpret-core==0.1.21
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
keras2onnx==1.6.0
kiwisolver==1.2.0
liac-arff==2.4.0
lightgbm==2.3.1
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
msal==1.4.3
msal-extensions==0.1.3
msrest==0.6.18
msrestazure==0.6.4
nbconvert==5.6.1
nbformat==5.0.7
ndg-httpsclient==0.5.1
nimbusml==1.7.1
notebook==6.1.1
numpy==1.19.1
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.0.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
pmdarima==1.1.1
portalocker==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.7.2   
py-cpuinfo==5.0.0
pyarrow==0.17.1
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
#pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
#pywin32
#pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
shap==0.34.0
six==1.15.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.9.0
statsmodels==0.10.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
#wincertstore==0.2
zipp==3.1.0
